+++ Source/iBMSUnreal/Public/Chart.h
@@ -26,8 +26,8 @@ class IBMSUNREAL_API FChartMeta {
 	FString SubTitle = "";
 	int Rank = 3;
 	double Total = 100;
-	long PlayLength = 0; // Timing of the last playable note, in microseconds
-	long TotalLength = 0; // Timing of the last timeline(including background note, bga change note, invisible note, ...), in microseconds
+	long long PlayLength = 0; // Timing of the last playable note, in microseconds
+	long long TotalLength = 0; // Timing of the last timeline(including background note, bga change note, invisible note, ...), in microseconds
 	FString Banner;
 	FString StageFile;
 	FString BackBmp;
